Bypass duplicate reference strain detected
Augur align detects the reference strain in the reference file and the curated dataset, and throws a "duplicate strain error"

`Duplicate strains of "KM822127" detected`

Usually I bypass this using `augur align --remove-reference` but this error is still showing up. Ergo, adding a postfix to the reference IDs to bypass error.